+++ zephyr/CMakeLists.txt
@@ -426,9 +426,19 @@ zephyr_library_sources_ifdef(CONFIG_COMP_FIR
 	${SOF_MATH_PATH}/fir_hifi3.c
 )
 
-zephyr_library_sources_ifdef(CONFIG_COMP_IIR
-	${SOF_AUDIO_PATH}/eq_iir/eq_iir.c
-)
+if(CONFIG_IPC_MAJOR_3)
+	zephyr_library_sources_ifdef(CONFIG_COMP_IIR
+		${SOF_AUDIO_PATH}/eq_iir/eq_iir.c
+		${SOF_AUDIO_PATH}/eq_iir/eq_iir_ipc3.c
+		${SOF_AUDIO_PATH}/eq_iir/eq_iir_generic.c
+	)
+elseif(CONFIG_IPC_MAJOR_4)
+	zephyr_library_sources_ifdef(CONFIG_COMP_IIR
+		${SOF_AUDIO_PATH}/eq_iir/eq_iir.c
+		${SOF_AUDIO_PATH}/eq_iir/eq_iir_ipc4.c
+		${SOF_AUDIO_PATH}/eq_iir/eq_iir_generic.c
+	)
+endif()
 
 zephyr_library_sources_ifdef(CONFIG_MATH_IIR_DF1
 	${SOF_MATH_PATH}/iir_df1_generic.c
